testing setup

+++ R/SampleDB_Setup.R
@@ -27,10 +27,14 @@ SampleDB_Setup <- function(){
   #https://stackoverflow.com/questions/1518729/change-sqlite-database-mode-to-read-write
   
   #make directory "/var/lib/sampleDB/"
-  system(paste("sudo -kS mkdir \"/var/lib/sampleDB\""), input = password)
+  system("sudo -kS mkdir /var/lib/sampleDB/", input = password) %>% suppressMessages()
+  system("sudo -kS mkdir /var/lib/sampleDB/backups", input = password) %>% suppressMessages()
+  system("sudo -kS mkdir /var/lib/sampleDB/move_files", input = password) %>% suppressMessages()
+  system("sudo -kS mkdir /var/lib/sampleDB/upload_files", input = password) %>% suppressMessages()
+  # stopifnot("" = dir.exists("/var/lib/sampleDB/") && dir.exists("/var/lib/sampleDB/backups/") && dir.exists("/var/lib/sampleDB/move_files") && dir.exists("/var/lib/sampleDB/upload_files"))
   
   #write sampledb .sqlite file to `var/lib/sampleDB/`
-  path <- "var/lib/sampleDB/"
+  path <- "/var/lib/sampleDB/"
   sqlite_file <- system.file("extdata", "sampledb_database.sqlite", package = "sampleDB")
   system(paste("sudo -kS cp", sqlite_file, path), input = password)
   message("Moved a SampleDB SQLite template file to \"/var/lib/sampleDB/sampledb_database.sqlite\"")
@@ -41,19 +45,32 @@ SampleDB_Setup <- function(){
   message("Added variable \"SDB_PATH\" to \"/etc/R/Renviron.site\"")
   
   #set up rshiny server
-  system("sudo -kS ln -s /usr/lib/R/site-library/sampleDB/sampleDB /srv/shiny-server", input = password)
+  # check if app was installed for user or for system
+  user_lib <- paste0(.libPaths()[1],"/sampleDB/sampleDB")
+  sys_lib <- paste0(.libPaths()[3],"/sampleDB/sampleDB")
+  if(dir.exists(sys_lib) && dir.exists(user_lib)){
+    system("sudo -kS ln -s /usr/lib/R/site-library/sampleDB/sampleDB /srv/shiny-server", input = password)
+  }else if(dir.exists(sys_lib)){
+    system("sudo -kS ln -s /usr/lib/R/site-library/sampleDB/sampleDB /srv/shiny-server", input = password)
+  }else if(dir.exists(user_lib)){
+    system(paste0("sudo -kS ln -s ", paste0(.libPaths()[1],"/sampleDB/sampleDB"), " /srv/shiny-server"), input = password)
+  }else{
+    message("Package needs to be installed in .libPaths()[1] or .libPaths()[3]")
+  }
+  # else 
+  # /usr/lib/R/site-library/sampleDB/sampleDB
+  # system("sudo -kS ln -s /usr/lib/R/site-library/sampleDB/sampleDB /srv/shiny-server", input = password) %>% suppressMessages()
   ip <- system("hostname -I | awk '{print $1}' 2>&1", intern = TRUE)
   url <- paste0("http://", ip, ":3838/sampleDB/")
   message(paste("Your SampleDB app is now available live at", url))
+  message("The sampleDB database lives at /var/lib/sampleDB/sampledb_database.sqlite")
+  message("SampleDB backups live at /var/lib/sampleDB/backups/")
   
   #setup backup system -- add a bash file (that backs up the database) to the app's extdata
   #ALL DIRS NEED TO BE CHMOD 777?
-  system(paste("sudo -kS mkdir \"/var/lib/sampleDB/backups\""), input = password)
-  system(paste("sudo -kS mkdir \"/var/lib/sampleDB/upload_files\""), input = password)
-  system(paste("sudo -kS mkdir \"/var/lib/sampleDB/move_files\""), input = password)
   #copy backup generator back script to /bin
-  sqlite_file <- system.file("extdata", "sampleDB_backup_generator.sh", package = "sampleDB")
-  path <- "/bin/"
-  system(paste("sudo -kS cp", sqlite_file, path), input = password)
+  # sqlite_file <- system.file("extdata", "sampleDB_backup_generator.sh", package = "sampleDB")
+  # path <- "/bin/"
+  # system(paste("sudo -kS cp", sqlite_file, path), input = password)
   
 }
